Add files via upload
Major changes to the library, use of gsl for matrix and vector manipulation
New functions:
- Convert from float* to gsl_matrix 
- Moore-Penrose pseudo inverse
- LMS sine regression

Several improvements have to be done wrt. memory use and freeing, as well as file format and the use of GSL in the main function.

Next steps are, however, to implement a configuration file for the acquisition, physically implement the MUX and improve a function to review acquisitions. Plotting is the next level.